chore: rename files to lower case
rename Statement.py and Config.py to be more in line with the
rest of the files in this project.
